export html template function to separate js file
