+++ src/make-html-template.js
@@ -1,0 +1,8 @@
+function makeHTMLtemplate(imgObject) {
+  const template = document.createElement('template');
+  const html = /*html*/`<li><h2>${imgObject.title}</h2><img src="${imgObject.url}"></li>`;
+  template.innerHTML = html;
+  const dom = template.content;
+  return dom;
+}
+export default makeHTMLtemplate;
